\name{pbdPROF-package}
\alias{pbdPROF-package}
\docType{package}
\title{MPI Profiling Tools}
\description{
  This package contains several libraries for MPI profiling, as well as some 
  R tools for parsing and analyzing their outputs.
}
\details{
\tabular{ll}{
  Package: \tab pbdPROF\cr
  Type: \tab Package\cr
  License: \tab MPL\cr
  LazyLoad: \tab yes\cr
}
  This package requires an MPI library (OpenMPI, MPICH2, or LAM/MPI).
}
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Gaurav Sehrawat, Wei-Chen Chen, Drew Schmidt, Pragneshkumar Patel, and
  George Ostrouchov.
}
\keyword{Package}
