\name{Profilers}
\alias{Profilers}

\title{Profilers}
\description{Profilers Directly Supported by pbdPROF}

\section{Details}{
  Currently, the \pkg{fpmpi} library is fully supported; a version of this 
  library is bundled with the source of the pbdPROF package (see package 
  vignette for more details).  Additional libraries can easily be linked with 
  pbdPROF, but these are not yet fully supported.  The \pkg{mpiP} and 
  \pkg{TAU} profilers are expected to be fully supported by the conclusion of 
  Google Summer of Code (~September 30 2013).
}
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}

  \pkg{fpmpi} website:
  \url{http://www.mcs.anl.gov/research/projects/fpmpi/WWW/}

  \pkg{mpiP} website:
  \url{http://mpip.sourceforge.net/}

  \pkg{TAU} website:
  \url{http://www.cs.uoregon.edu/research/tau/home.php}
}

