\name{Example datasets}
\alias{fpmpi.allreduce}
\alias{fpmpi.masterslavePI}
\alias{fpmpi.svd}
\alias{mpip.allreduce}
\alias{mpip.masterslavePI}
\alias{mpip.svd}
\docType{data}
\title{Small example datasets}
\description{
  These datasets are small for test operations and functions.
}
\format{
  Each dataset contains information profiled by \pkg{fpmpi} or \pkg{mpip}
  for \code{pbdMPI/demo/allreduce.r}, \code{pbdDMAT/demo/svd.r}, and
  \code{Rmpi/demo/masterslavePI.r}.
}
\details{
  The example datasets are profiled in 2 processors.
}
\references{
  fpmpi website:
  \url{http://??}

  mpiP website:
  \url{http://??}

  TAU website:
  \url{http://??}

  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Gaurav Sehrawat, Wei-Chen Chen, Drew Schmidt, Pragneshkumar Patel, and
  George Ostrouchov.
}
\keyword{data}
