\name{Print}
\docType{methods}

\alias{print-method}
\alias{print,prof-method}
\alias{print}

\alias{show-method}
\alias{show,prof-method}
\alias{show}

\title{ Printing }
\description{
  Print and show methods for \code{prof} class objects.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "prof")}}{}
}
}
\usage{
  \S4method{print}{prof}(x, ...)
  \S4method{show}{prof}(object)
}
\arguments{
  \item{x, object}{\code{prof} class object}
  \item{...}{extra arguments}
}
\seealso{
  \code{
    \link{prof-class}, 
    \link{read.prof}
  }
}
\keyword{Methods}

