% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdPROF-package.r
\docType{package}
\name{pbdPROF-package}
\alias{pbdPROF}
\alias{pbdPROF-package}
\title{MPI Profiling Tools}
\description{
This package contains several libraries for profiling MPI codes, as well as
some R tools for parsing, analyzing, and plotting their outputs.
}
\details{
\tabular{ll}{ Package: \tab pbdPROF\cr Type: \tab Package\cr License: \tab
MPL\cr LazyLoad: \tab yes\cr } This package requires an MPI library
(OpenMPI, MPICH2, or LAM/MPI).
}
\examples{

\dontrun{
demo(allreduce.fpmpi, "pbdPROF")
demo(svd.fpmpi, "pbdPROF")
demo(masterslavePI.fpmpi, "pbdPROF")
demo(allreduce.mpip, "pbdPROF")
demo(svd.mpip, "pbdPROF")
demo(masterslavePI.mpip, "pbdPROF")
}

}
\author{
Wei-Chen Chen, Drew Schmidt, Gaurav Sehrawat, Pragneshkumar Patel,
and George Ostrouchov.
}
\references{
Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\keyword{Package}

