% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.r
\docType{class}
\name{prof-class}
\alias{fpmpi-class}
\alias{mpip-class}
\alias{prof}
\alias{prof-class}
\alias{rawprof-class}
\alias{tau-class}
\title{Class prof}
\description{
Class for Profiler Output
}
\section{Slots}{

\describe{
\item{\code{profiler}}{The type of profiler used (e.g. fpmpi, mpiP). Stored as a character
vector.}

\item{\code{raw}}{The raw (non-processed) profiler output.  Storage is basically a
character vector, but set as a virtual class \code{rawprof}.}

\item{\code{parsed}}{A dataframe containing the processed version of the raw data.}
}}
\seealso{
\code{\link{read.prof}}
}
\keyword{Classes}

