% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\docType{methods}
\name{prof-print}
\alias{print,prof-method}
\alias{prof-print}
\alias{show,prof-method}
\title{Printing}
\usage{
\S4method{print}{prof}(x, ...)

\S4method{show}{prof}(object)
}
\arguments{
\item{x, object}{A \code{prof} class object}

\item{...}{extra arguments}
}
\description{
Print and show methods for \code{prof} class objects.
}
\seealso{
\code{ \link{prof-class}, \link{read.prof} }
}
\keyword{Methods}

