\name{Example datasets}
\alias{fpmpi.allreduce}
\alias{fpmpi.masterslavePI}
\alias{fpmpi.svd}
\alias{mpip.allreduce}
\alias{mpip.masterslavePI}
\alias{mpip.svd}
\docType{data}
\title{Small example datasets}
\description{
  These datasets are small for test operations and functions.These datasets will be used in demo
  part for showing parsing and plotting.They are located in exdata directory of installated package.
}
\format{
  Each dataset contains information profiled by \pkg{fpmpi} or \pkg{mpip}
  for \code{pbdMPI/demo/allreduce.r}, \code{pbdDMAT/demo/svd.r}, and
  \code{Rmpi/demo/masterslavePI.r}.
}
\details{
  The example datasets are profiled in 2 processors.
}
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}

  fpmpi website:
  \url{http://www.mcs.anl.gov/research/projects/fpmpi/WWW/}

  mpiP website:
  \url{http://mpip.sourceforge.net/}

  TAU website:
  \url{http://www.cs.uoregon.edu/research/tau/home.php}
}
\keyword{data}
